fix: overview rerendering
